[Flutter] Add JavaScript sourcemap upload
This adds the ability to upload JavaScript sourcemaps as part of the `flutter-symbols upload` command.

Manual upload of Flutter source maps was already possible using the `sourcemaps upload` command, this just integrates it with the `flutter-symbols` command, allows automatic detection of version (if possible) and allows you to use one command over two.