[CIVIS-9031] Deprecate 'metrics' naming in favour of 'measures' in junit cmd
This patch adds new `--measures` and `--report-measures` options to the
junit command. These options have the same functionality as the
`--metrics` and `--report-metrics` options. When using the old `metrics`
options, we also print a deprecation warning.

Aditionally, we replace the mentions of "metrics" with "measures" in
the docs.

Context: https://docs.google.com/document/d/1X9-0eiiNaDsfLMKlVYsGdQs4RDKGZx9jEqLX4NEPAxY