Add elf symbols upload command (#1275)
This adds elf-symbols upload command that scans input location for any elf files, extracts debug information from them and uploads it to Datadog intake with proper information.
Elf debug information will be used by profiling backend to symbolicate native profiles.

This command is intended to run on a system with `objcopy`. It accepts either a single file or a directory as input.
If input is a directory, it is walked recursively, looking for all files that have ELF magic number.

Only elf files that match the following conditions are considered:

* x86_64 / aarch64 or arm architecture
* ELF type EXEC or DYN (executables or shared libraries)
* have a .debug_info section or .symtab section (symbol table)

I could not find a JS library for ELF parsing (there are a few ones, but without much use and maintenance), therefore I implemented a basic ELF parser that parses ELF file header, program headers and sections headers.

Once the ELF file list with the associated metadata is built, we check for duplicate build ids since we don't want to upload debug information associated to the same build id multiple times. To deduplicate build ids, we prefer files having debug infos over files that don't, and files that have symbol tables over those that don't, otherwise we keep the first occurence.

Both GNU and Go buildids if present are sent as metadata, also compute a file hash as an alternative to the buildid.
File hash is sent only if file has code.

Finally for each ELF file, debug info is extracted in a temporary directory with `objcopy  --only-keep-debug` command, we also remove .gdb_index section (`--remove-section=.gdb_index`) since it can be quite big and is not useful for symbolication and then debug info file is uploaded to Datadog.

Once a file is determined to be an ELF file, most failures later on in the process result in warnings, except when input is a single file where any failure is considered an error.