(DoD for ticket)make clear it's the same global config for all commands
This is the the commit for the 1/2 of the actions required for my ticket.
I'm explicitly calling out in the upload-application command that it can be the same Global Config file used for both commands.
Also renamed all the uses of a global config file i could see in the README to all use the same name for the file so it's less confusing when reading