[elf-symbols] Fix dynamic symbols extraction (#1713)
* elf-symbols: fix dynamic symbols extraction

This change fixes and simplifies the dynamic symbols extraction process.
Previous implementation was dumping `.dynsym/.dynstr` sections content
in temporary files and then removing those sections before adding them back
to the debug info file (because `--only-keep-debug` was setting their
size to 0).

The new implementation simply sets the sections flags to `alloc,readonly,contents`
(which is the default for those sections) and this seems to prevent
`--only-keep-debug` from removing their content.

* Fix dynamic symbols extraction for symbolic

Symbolic library requires dynamic section/segment and some other sections
(hash sections, version sections, relocation sections) to be present in the
in the debug info file to be able to create a symcache.